+++ src/kbforge/core/__init__.py
@@ -5,5 +5,6 @@ package are thin wrappers that call into this package.
 """
 
 from .pipeline import run
+from .site import build_site
 
-__all__ = ["run"]
+__all__ = ["run", "build_site"]